+++ CHANGELOG.md
@@ -41,6 +41,10 @@
 * Bugfix:
    * App Release Process
       * Fixed Conformity Document Deletion is not backend connected issue
+* Shared Component
+   * New About page component added
+   * About page created
+   * Removed all Third party related contents
 * Updates
    * Service release process
       * SERVICE_LEADIMAGE supported.